Gzip more type of files and force some extension to text/plain
This also creates the package scriptworker.artifacts
